feat: factor out faq vm
