+++ App/UI/FAQ/FaqView.swift
@@ -16,86 +16,6 @@ import Foundation
 import Models
 import Tempura
 
-struct FaqVM: ViewModelWithLocalState {
-  /// The list of FAQ to show
-  let faqs: [FAQ]
-  /// Whether the view is presented modally. This will change the back/close button visibility.
-  let isPresentedModally: Bool
-  /// Whether the header is visible in the view. The header is shown only when the content is scrolled.
-  let isHeaderVisible: Bool
-  /// Whether the user is actively searching through the search bar.
-  let isSearching: Bool
-  /// The currently searching string.
-  let searchFilter: String
-
-  func shouldUpdateHeader(oldModel: FaqVM?) -> Bool {
-    return self.isHeaderVisible != oldModel?.isHeaderVisible
-  }
-
-  func shouldUpdateSearchStatus(oldModel: FaqVM?) -> Bool {
-    return self.isSearching != oldModel?.isSearching
-  }
-
-  func shouldReloadCollection(oldModel: FaqVM?) -> Bool {
-    guard let oldModel = oldModel else {
-      return true
-    }
-
-    if self.faqs != oldModel.faqs {
-      return true
-    }
-
-    if self.searchFilter != oldModel.searchFilter {
-      return true
-    }
-
-    return false
-  }
-
-  func cellModel(for indexPath: IndexPath) -> ViewModel? {
-    guard let faq = self.faqs[safe: indexPath.item] else {
-      return nil
-    }
-    return FaqCellVM(faq: faq, searchFilter: self.searchFilter)
-  }
-
-  var searchBarVM: SearchBarVM { SearchBarVM(isSearching: self.isSearching) }
-  var shouldShowSeparator: Bool { !self.isHeaderVisible }
-  var shouldShowTitle: Bool { !self.isSearching }
-  var shouldShowBackButton: Bool { !self.isPresentedModally && !self.isSearching }
-  var shouldShowCloseButton: Bool { self.isPresentedModally && !self.isSearching }
-}
-
-extension FaqVM {
-  init?(state: AppState?, localState: FAQLS) {
-    guard
-      let state = state,
-      let faqs = state.faq.faqs(for: state.environment.userLanguage)
-      else {
-        return nil
-    }
-
-    if localState.searchFilter.isEmpty {
-      self.faqs = faqs
-    } else {
-      let exactMatching = faqs.filter { $0.title.lowercased().contains(localState.searchFilter.lowercased()) }
-      let fuzzyMatching = faqs.filter {
-        !exactMatching.contains($0) &&
-          (
-            $0.title.lowercased().fuzzyContains(localState.searchFilter.lowercased()) ||
-              $0.content.lowercased().contains(localState.searchFilter.lowercased())
-          )
-      }
-      self.faqs = exactMatching + fuzzyMatching
-    }
-
-    self.isPresentedModally = localState.isPresentedModally
-    self.isHeaderVisible = localState.isHeaderVisible
-    self.isSearching = localState.isSearching
-    self.searchFilter = localState.searchFilter
-  }
-}
-
 // MARK: - View
 
 class FaqView: UIView, ViewControllerModellableView {
